[Bug Fix] Adds a Check For MTCP Include Directory (#3)
In 4ce119b0, DPDK has been modified for MTCP and generated information regarding cflags and ldflags to text files.  The build process did not verify whether the `include` directory that these text files are generated in exists.

This bug fix addresses this by adding a conditional to check and create the directory if it does not exist.

Commit log:

* Bugfix for cflags.txt file generation

* Fixed the folder creation