Allow igb_uio Module to Set Up Network Adapters (#2)
This addition modifies the igb_uio kernel module such that NIC ports that are being managed by it can be setup as network adapters that are visible to the Linux kernel (i.e. running `ifconfig`).

This change is necessary for running complex NFs such as MTCP, a userspace TCP stack.  To achieve this, the source code for the module is amended such that it sets up the `struct net_device` and `struct net_adapter` data structures for the system to enable adapters.

We also provide a script to enable interfaces such that the Linux kernel can see them (i.e. running ifconfig).

The flags used to compile the igb_uio module are exported to a text file so other applications can reference them.  This is needed for complex NFs such as MTCP.

Commit log:

* dpdk changes for mtcp stack

* flags for dpdk

* Removed cflags, ldflags txt

 - Added commands to generate cflags, ldflags

* Changed setup_iface script for dpdk interface

- Changed args to include interface name, ip, netmask, up/down

* Made requested changes to iface script

* Modified iface script to close interface

Conflicts:
	lib/librte_eal/linuxapp/igb_uio/igb_uio.c